fix: Remove unused cryptiles dependency
When installing `graphql-playground-react` there are warnings being thrown that are related to an unused dependency, `cryptiles`. Since we're not using this anymore, there is reason to keep it in the `package.json`.

```
npm WARN deprecated cryptiles@4.1.2: This version has been deprecated in accordance with the hapi support policy (hapi.im/support). Please upgrade to the latest version to get the best features, bug fixes, and security patches. If you are unable to upgrade at this time, paid support is available for older versions (hapi.im/commercial).
npm WARN deprecated boom@7.3.0: This module has moved and is now available at @hapi/boom. Please update your dependencies as this version is no longer maintained an may contain bugs and security issues.
```